backend/requirements.txt
flask
flask-cors
pdfplumber
pytesseract
Pillow
opencv-python
numpy
